Refactored JavaCommander, introduced ArgumentsParser and Tokenizer.
